Removed novelupdates as template, keeping it as source only
